phase 3.6: ship sim2real loop module
